Copied polishedVersion to Carson folder.

+++ src/Carson/Drawable.java
@@ -4,8 +4,8 @@ package Carson;
  * Drawable is an abstract class which should be implemented by classes that can be on screen, either directly or from the inheritance chain
  */
 public abstract class Drawable {
-	int drawX;
-	int drawY;
+	double drawX;
+	double drawY;
 	double x;
 	double y;
 	
@@ -15,8 +15,8 @@ public abstract class Drawable {
 	}
 	
 	public void updateDrawXY(double x, double y) {
-		this.drawX = (int) x;
-		this.drawY = (int) y;
+		this.drawX = x;
+		this.drawY = y;
 	}
 	
 }
